frida-proxy/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

click==8.1.7
frida==16.4.8
grpcio==1.65.5
loguru==0.7.2
